SecLLM/pyproject.toml
[tool.poetry]
name = "secllm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.54.3"
pyyaml = "^6.0.2"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"
seaborn = "^0.13.2"
nltk = "^3.9.1"
rich = "^13.9.4"
anthropic = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
